HW-6 BaseViewModel 외부로 이동
