+++ dlwls5201/MyArchitecture/app/src/main/java/study/architecture/myarchitecture/BaseViewHolder.kt
@@ -1,0 +1,19 @@
+package study.architecture.myarchitecture
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.recyclerview.widget.RecyclerView
+
+abstract class BaseViewHolder<B : ViewDataBinding>(
+    @LayoutRes layoutRes: Int,
+    parent: ViewGroup?
+) : RecyclerView.ViewHolder(
+    LayoutInflater.from(parent?.context)
+        .inflate(layoutRes, parent, false)
+) {
+
+    val binding: B = DataBindingUtil.bind(itemView)!!
+}
