+++ dlwls5201/MyArchitecture/app/src/main/java/study/architecture/myarchitecture/ui/tickerlist/TickerAdapter.kt
@@ -1,11 +1,8 @@
 package study.architecture.myarchitecture.ui.tickerlist
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
+import study.architecture.myarchitecture.BaseViewHolder
 import study.architecture.myarchitecture.R
 import study.architecture.myarchitecture.databinding.ItemTickerBinding
 import study.architecture.myarchitecture.ui.model.TickerItem
@@ -58,17 +55,6 @@ class TickerAdapter(
                 executePendingBindings()
             }
         }
-    }
-
-    abstract class BaseViewHolder<B : ViewDataBinding>(
-        @LayoutRes layoutRes: Int,
-        parent: ViewGroup?
-    ) : RecyclerView.ViewHolder(
-        LayoutInflater.from(parent?.context)
-            .inflate(layoutRes, parent, false)
-    ) {
-
-        val binding: B = DataBindingUtil.bind(itemView)!!
     }
 
 }
